PID Improvements (untested) Dimmer tuning for 60Hz (untested)

+++ avr/src/app/TimedDrying.cpp
@@ -203,15 +203,15 @@ void TimedDrying::changeSelected(int8_t amount) {
     switch (selection) {
         case TEMPERATURE:
             setTemperature += amount;
-            setTemperature = Math::max(0, Math::min(70, setTemperature));
+            setTemperature = Math::max((int8_t)0, Math::min((int8_t)70, setTemperature));
             break;
         case MINUTES:
             setMinutes += amount;
-            setMinutes = Math::max(0, Math::min(120, setMinutes));
+            setMinutes = Math::max((int8_t)0, Math::min((int8_t)120, setMinutes));
             break;
         case SECONDS:
             setSeconds += amount;
-            setSeconds = Math::max(0, Math::min(59, setSeconds));
+            setSeconds = Math::max((int8_t)0, Math::min((int8_t)59, setSeconds));
             break;
         default:
             break;
